+++ Changelog.md
@@ -1,5 +1,6 @@
 ## 0.0.64
 + Added `Parameter.DepthMap` annotation method to specify depth map parameters in depth estimation predictors.
++ Fixed `muna.beta.openai.embeddings.create` throwing error due to unhashable type.
 
 ## 0.0.63
 + Added `muna.beta.openai.embeddings.create` method for using text embedding predictors via the OpenAI API.